refactor(certificate): clean dead code and improve variable naming
  - Remove unused mockup file useRegulatoryData.js (91 lines of dead code)
  - Remove unused constants THRESHOLD_LABEL_MAPPING and TAB_LABELS
  - Fix filename typo: certifiationCriterias.js → certificationCriterias.js
  - Update import reference in CertificationCriteriaGlobalResult.js
  - Add placeholder implementation for handleDownloadCertificate function
  - Improve variable naming for better code readability:
    * conditionalTitle → titleByComplianceStatus
    * statusClass → complianceStatusClass
    * matchingData → thresholdDetail
    * thresholdConfigs → categoryThresholds
