+++ public/index.html
@@ -96,7 +96,7 @@
 
     <link
       rel="stylesheet"
-      href="%PUBLIC_URL%/dsfr/utility/icons/icons.min.css?hash=2be755ac"
+      href="%PUBLIC_URL%/dsfr/utility/icons/icons.min.css?hash=1d84e762"
     />
     <link rel="stylesheet" href="%PUBLIC_URL%/dsfr/dsfr.min.css?v=1.13.2" />
     <title>Mobilic</title>
